feat: set CU budget and per CU fees to more reliably process commit txs (#36)
## Summary

Adding Compute Budget instructions to set max compute units and
configurable per CU price.

## Details

We had trouble getting the commit transaction processed on mainnet.
After figuring out a large
enough CU budget and a compute unit price that reliably let them through
we added instructions
to the commit transaction to set those values.

The CU budget is hardcoded for now and the compute unit price can be
configured defaulting to
the min value we saw to get reliable processing.

Additionally the committer now logs before and after the commit
transaction to make it easier
to detect a transaction that is stuck.

## Unrelated

Updated the workflows to only run on push or PR to master.