chore: improve memory management + CPU perf for status cache (#37)
## Summary

Fix status cache memory growth by purging ole entries regularly.

Adding status cache optimized for transaction signature lookup to
improve performance.

## Details

The status cache was never purged so far and thus grew in memory. Now we
regularly clean it and
only retain items that were added after a specific slot.
The amount of recent slots we keep is derived from the millis per slot
in order to roughly keep
a specific time window in the cache.

When subscribing to signature status we looked up if the signature is
already in our bank
(status cache). With a large cache that caused a lot of overhead which
was partially fixed by
keeping the cache small.
To optimize it even further I added a transaction status cache keyed by
slot (the others are
keyed by blockhash) and added it to the items that are purged regularly.

As a result this lookup that could take close to 50% of the CPU time and
more depending on the
size of the status cache now does not even show up in the flamegraph
anymore.

---------

Co-authored-by: crypto-vincent <106254129+crypto-vincent@users.noreply.github.com>