feat: simpler cache implementation in geyser plugin (#39)
## Summary

Replacing the _stretto_ cache inside the Geyser plugin with a custom
circular buffer + hashmap
based implementation.

## Details

The Geyser plugin needs to cache recent transaction and account updates
in case a subscription
comes in late and misses the event it is actually interested in. Due to
our validator being
that much faster this happens quite often. Therefore we serve recent
events from a cache to
avoid that race condition.

Up until now we were using the _stretto_ cache crate but it turned out
too complex to configure
and caused a bottleneck in the transaction execution pipeline due to the
overhead when adding
an item to the cache. We observed this overhead in a flamegraph.

With this much simpler low overhead cache the behavior is the same and
the flamegraph shows no
more time spent inserting items into the cache.

Additionally simplifying this will make further memory investigations
easier.