chore: fix geyser memory issue (#40)
## Summary

Removing the largest part source of memory increase over time.

## Details

Inside the Geyser code a message status cleanup function was not being
run due to it depending
on the `Finalized` commitment.

This was fixed along with some improvements and abstractions in the
related code.

## Status

I ran some benchmarks to confirm that the cleanup code does not affect
CPU perf.
The largest pieces in the transaction processing pipeline are now sig
verify and the actual
processing and result storing of each transaction.

With lots of transactions running memory does increase and went to about
6GB after 4M
transactions.
Without any transaction load memory does not increas (verified with
Instruments).

There might still be some places where state is accumulated that needs
to be cleaned up
regularly but for now this is a large improvement over the previous
status.

We should address those in follow up investigations, but I'd say that
the current rate of
memory growth is acceptable to deploy this remotely.

---------

Co-authored-by: crypto-vincent <106254129+crypto-vincent@users.noreply.github.com>