feat: manual commit triggers via CPI (#68)
## Summary

Fixes manual account commit triggers such that they can be invoked as
part of a CPI from inside
another program and speeds all trigger transactions up by not waiting
for the commit to
actually complete.

Adds integration tests for direct and CPI account commit triggers.

CLOSES: #62

## Details

Account commits used tokio channels to communicate the commit intent and
the initial result
including the signature. This caused issues since we ended up blocking
on a tokio task which
crashed the thread. We now use crossbeam channel for this communication.

Initially we would wait until the commit transaction on chain completed
and blocked transaction
processing on that thread. Now we only create the transaction and
immediately callback with the
signature. The transaction can now complete and the on-chain commit is
completed asynchronously.
The ephemeral transaction logs either of the below to indicate that the
commit happened with
the signature or did not since the account state didn't change:

```
TriggerCommit: 93S8QLdefznMhisJVGvHx8DmxZZd7i3cERAKQTEio1w8
```

```
NotCommitted
```

## Tests

We had not integration tests in place at all so I set up the first
iteration on this as part of
the test-programs workspace to which I also added the trigger commit
client + CPI program.
These tests start the validator and then attempt to trigger commits
either directly or via CPI.

They now run as part of the main test suite.
They use devnet to send those commits to, thus we need to make sure it
is available in CI and
doesn't cause any issues, i.e. due to congestion.

## Unrelated

Adding `Cargo.lock` files to work around yanked crates and other dep
issues.

## Remaining

- once we add more integration tests, pull out a reusable framework for
this using a similar
approach

Some clients want to commit state that is changed as part of a previous
instruction in the same
transaction which is not addressed at this point for the following
reasons:

- when committing we pull the _current bank state_ not the state that
will be persisted _after_
  the transaction completes
- even if we could somehow communicate the _updated state_ when
committing this would introduce
a security whole as now I could commit a state that might not be
accepted since the
  instruction setting it would fail

Some solutions would be:

- allow client to schedule a transaction to run right after a
transaction or with the next
  crank loop
- this scheduling could be done from inside the instruction itself via a
new CPI call that we
  would have to introduce