feat: enable multi-instances on same machine (#74)
## Summary

Adding the below config value in order to allow overriding the config
port at which the Geyser
GRPC service listens.

```toml
[geyser_grpc]
addr = "127.0.0.1"
port = 11000
```

This is needed in order to run multiple instances of the validator on
the same machine.

## Unrelated

Added a way to detect if devnet is up and skip some tests that depend on
it if it is not.