feat: add compute budget program to builtins (#81)
## Summary

Adding compute budget program to loaded builtins and improving error
messages for cases where
executable data of a program cannot be found.

## Details

### Issue Encountered

We ran into an issue which resulted in the following error for a
customer:

```
MutatorError(RpcClientError(Error {
  request: None, kind: RpcError(ForUser(
    \"AccountNotFound: pubkey=7svwMk17uFMeZ8vwUwVgnuVpmCje1YuD9RrAsz2pDYQj\"))
})). \nCatch the `SendTransactionError and call getLogs()
```

This was hard to trace as we didn't understand the significance of the
`7svwMk17uFMeZ8vwUwVgnuVpmCje1YuD9RrAsz2pDYQj` account.

It is the executable data account for the compute budget program if it
wasn't builtin.

### Improved Error Message

The error message for such cases now has been improved to look like the
below instead:

```
"MutatorError(FailedToCloneProgramExecutableDataAccount(
  \"ComputeBudget111111111111111111111111111111\", Error { request: None, kind: RpcError(ForUser(
    \"AccountNotFound: pubkey=7svwMk17uFMeZ8vwUwVgnuVpmCje1YuD9RrAsz2pDYQj\"))
}))"}
```

which will help in realizing cases where we don't support builtins much
quicker.

### Fix

We have added the compute budget program to the loaded builtins so that
it is always available
and will not be attempted to be cloned anymore.