feat: simpler schedule commit API (#115)
## Summary

Simplify schedule Commit API.

## Details

1. No longer requiring validator id since we removed cost transfer
2. No longer requiring `invoke_signed` of the CPI instruction and thus
removing the need for
seeds

### 2. CPI invoke Detection over PDA signers

Simplification `2.` entails that we can no longer rely on signer checks
of the PDAs we are
committing in order to ensure that the request originated from a CPI of
the program that owns
those PDAs.

Instead we detect the _parent_ instruction that invoked us and get its
`proram_id`.
That is all the security that we require in order to ensure that the
program itself is
responsible for requesting the scheduled commit via CPI.

We need that security to guarantee that accounts that need to be
committed atomically are
committed like that (the program gets to ensure that).

#### Security

In order to verify that no attack vector exists, i.e. the validator
being tricked into thinking
that a CPI came from a specific program when it didn't I diagnosed in
high detail what
information the `TransactionContext` provides us. I documented this

[here](https://github.com/magicblock-labs/magicblock-validator/discussions/108#discussion-6999643)
as well as recorded data that this PR includes inside the
`./test-programs/schedulecommit/security/tests/results/` directory.
There the _debug_ output of
those contexts in each case was included.
Additionally for each case we now run a test that checks that we fail
incorrectly attempted
commit schedules.

## Remaining

While working on this I encountered a potential issue and [recorded it

here](https://github.com/magicblock-labs/magicblock-validator/discussions/108#discussioncomment-10211631).

If preferred I can convert this into a separate issue.