chore: schedule commits are more resistant to timeouts (#126)
## Summary

https://github.com/magicblock-labs/magicblock-validator/pull/116 runs
into inconsistent test
outcomes due to timeouts/race conditions.

The below hardens the tests to deal with transactions taking way longer
to confirm than
expected as well as print more information in case of issues.

I noted that on this branch I don't see those timeouts, but see them
consistently on the
version of the above linked PR.

---------

Co-authored-by: crypto-vincent <vincent.brunet.us@gmail.com>