fix: invalid local delegation program causing ix test failures (#143)
## Summary

Using local delegation program for integration tests.

## Details

We cannot rely on the version on chain since our PRs may depend on newer
ones that aren't
deployed yet.

### Fixed Issues

When using the local program we ran into the following issues:

- when using Delegation program via CPI it errored with program not
loaded
- tracking this down I found the following section deep inside
`solana-program-runtime/src/loaded_programs.rs`:

```rust
#[cfg(all(not(target_os = "windows"), target_arch = "x86_64"))]
{
    let mut jit_compile_time = Measure::start("jit_compile_time");
    executable.jit_compile()?;
    jit_compile_time.stop();
    metrics.jit_compile_us = jit_compile_time.as_us();
}
```

The line `executable.jit_compile()?;` failed for the Delegation program.
However since this section only runs on x86_64 we only saw this issue on
our CI
and locally only on my Intel.

To fix this I reset the delegation program to commit `f01e716`,
recompiled with
`cargo build-sbf` and replaced the binary we're using.

Then I verified that we no longer error on the jit compiler line and
that the tests now pass.
To make this work in CI we need to clone the delegation program at that
ref until the commit+undelegate work is done since all previous
validator versions are not yet compatible with the latest delegation
program version.

### Logging

I adapted the logging for the integration tests to put out a ton more
information so that in the future we can debug them quicker in CI.