feat: integrate BPF loaded into clone/mutator pipeline (#135)
## Summary

When cloning programs, we need to apply special treatment to please the
BPFUpgradeLoader.
We need to run an `Upgrade` IX right after mutating the programs
accounts so that the execution cache picks up that the program has
changed and potentially updates behavior.

## Details

This PR introduce the use of the BPF upgrade loader's `Upgrade`
instruction after having dumped the AccountModification for each cloned
program. This require a small refactor of the mutator module.

Other implementation changes:
- Deduplicate the data structure: `AccountModification` (we used to have
2 copies)
- `AccountCloner.clone_account` now returns `Vec<Signature>` instead of
`Signature`
 - Removed the use of `str` in place of a `Pubkey`

## Notes

Deliberately not addressing thoses issues in this PR (Will address later
to minimize this current DIFF):
 - https://github.com/magicblock-labs/magicblock-validator/issues/136 
 - https://github.com/magicblock-labs/magicblock-validator/issues/138
 - https://github.com/magicblock-labs/magicblock-validator/issues/139

---------

Co-authored-by: Thorsten Lorenz <thlorenz@gmx.de>