feat: optional undelegate as part of commit (#146)
## Summary

When committing an account via the `ScheduleCommitAndUndelegate`
instruction the account is
committed and undelegated on chain and removed from the ephemeral
validator.

## Still Missing

- unit tests for accounts remover (if possible since in a mock context I
might not be allowed to drain accounts that the validator auth does not
own)

So even though the unit tests would be nice the integration tests cover
the happy path of accounts removal at least.
Still I'd like to add tests to ensure that this transaction fails if the
validator didn't sign for instance.
However this could be addressed in a follow up PR.

## Details

The steps for an account that is scheduled to be committed and
undelegated are as follows:

1. User sends transaction to schedule commit and undelegation
2. Transaction is processed
2.1. the account is marked as scheduled to be committed and undelegated
inside a global collection
2.2. the owner of the account is set to the delegation program which
makes it unusable for the
time being (except for reads), such that no following transaction can
write to it
3. In the next slot this account is processed:
3.1. A transaction to notify of processing is executed in the ephemeral
3.2. A transaction to commit and undelegate the account is sent to chain
3.3. We start listening for the outcome of 3.2 on a separate thread
4. The on-chain commit + undelegate transaction completes successfully
5. We request to remove the account from the ephemeral by registering it
with the accounts
remover hashset
6. At the next slot start we detect that accounts are requested to be
removed
6.1. We remove the account from the external accounts that are tracked
by the cloning pipeline,
aka the accounts manager
6.2. We execute a transaction that drains the account, thus removing it
from the bank, we also
remove it from the hashset of the accounts remover as part of that
transaction
7. If now the account is used by a transaction again the cloning
pipeline behaves as if we
never cloned it

[Design
Diagram](https://drive.google.com/file/d/1HZdopSv-Kt88epNT--2LpaBa20mqGBw-/view?usp=sharing)