hotfix: preventing not needed account commits draining our Sol (#147)
## Issue

Currently the commit transaction to update accounts on chain is sent
even if no accounts need
to be committed.
This is caused by the _frequent commit_ feature.

## Fix

I added a check that verifies that we actually are committing accounts
instead of sending a
basically empty transaction.

## Tests

In order to fix this quickly this PR includes no added tests, however I
will include one with a
follow up PR which also will be the first step in providing a reusable
integration test setup.