feat: restructured ix tests which can run with the solana-test-validator (#178)
## Summary

Reorganizing integration tests and enabling running chain instance with
solana test validator.

## Details

The test runner now supports running the devnet validator in two modes:

1. Using `solana-test-validator` with the accounts and programs
pre-configured (default)
2. Using the MagicBlock validator (enabled by setting
`FORCE_MAGIC_BLOCK_VALIDATOR` env var)

In order to not repeat configs we added a way to convert our config
tomls into arguments we
pass to the `solana-test-validator` command. Mainly to configure RPC
port and programs to load.

Additionally we always fund the authories we use with Luzid or our
Validator by loading the
accounts into the validator with 20 SOL.

By default CI tests now run with the solana-test-validator for the chain
instance and an action to install
the tooling was added.

However we are currently overriding that behavior as we discovered that
this surfaces more issues which we will have to triage and fix in a
separate PR.
This is tracked here
https://github.com/magicblock-labs/magicblock-validator/issues/183