feat: ledger initialization/config improvements (#214)
## Summary

This pull request removes an obsolete ledger option and adds a lock to
avoid multiple
validators for the same ledger.

## Details

The ability to provide an already initialized ledger conflicted with the
other options, i.e.
`reset` which would remove the ledger directory _after_ it was
initialized.
This was causing actual issues for me which were hard to debug and
resulted in no ledger data being persisted.

After this change I no longer run into this issue as the `reset` is
always applied _before_ the
ledger itself is initialized.

The ledger lock file implementation was more ore less copied from
Solana.

## Follow Up

Luzid needs to be adapted to provide a ledger path instead of an already
initialized ledger.
Additionally the implementation related to ledger lock file and cleaning
of the ledger
directory can now be removed from there.

CLOSES: #212 (adding last missing metric - ledger)