feat: restore cloned accounts during ledger replay (#248)
## Summary

Restoring accounts that were cloned into the validator and integration
test improvements.

After restoring accounts can not be used yet since the account
cloning/delegation metadata is
not restored at this point.

## Details

- cleaned up mutation account program and related modules
- added an account mod data persister which stores cloned data in the
ledger and loads it from
  there during ledger replay
- showed that delegated and then cloned accounts are restored correctly
- added another test showing that cloned accounts cannot be used yet
(disabled this for now,
but it could serve as a start for the remaining work)
- added a separate doc explaining the commit workflow and how it relates
to ledger replay
- adding global flag indicating if validator started up or is replaying
ledger
- used above flag to prevent failure of `SentCommit` instruction by
bypassing its logic which
is not useful during ledger replay
- ensuring that we don't try to load account mods from the ledger unless
we're replaying it at
  that point in time

## Tests

Many ledger restore specific integration tests were added and the test
context improved and
enhanced.
Additionally via the [cleanass](https://crates.io/crates/cleanass)
crate, the assertions are
now killing a running validator before they cause a panic in case an
assert fails.

- added delegation/commit related instructions to flexi counter
- refactored out a reusable scheduled commits tracker so we can easily
verify those all the way
  to chain in integration tests
- reorganized test runner and added another backup to ensure that
validators are properly
killed after each suite completes

## Unrelated

- put clock account update log on trace since it was spamming logs on
`info` making them
unusable during development

## Remaining

Currently accounts that were cloned into the validator are not usable
after the ledger replay.
This is due to the account metadata not being restored.

- hydrate metadata of accounts that are in the validator such that they
are usable after the
ledger replay is complete
- update the account mod ID to not start at zero, but at +1 from the
last ID used when the
ledger was replayed

We should verify the state of replayed commits that never were realized,
namely the
following:

- state of the `MagicContext1111...` account
- NOTE: the state of globally `ScheduledCommits` is already logged and
then cleared

We need to clear both of them before we start the validator after ledger
replay.

- the `MagicContext1111...` account is cleared after the scheduled
commits are moved to the
global, but we should need to verify this still