feat: hydrate bank with accounts for ledger replay + tool improvs (#273)
## Summary

This PR hydrates accounts from the storage included with the ledger.

This means they are first loaded into the bank, then the ledger is
replayed and finally
all accounts except the ones delegated to our validator are cloned from
the remote.

In order to test/triage this tools were improved and added.

I was able to fully replay a leger with ~10million slots and clone its
~800 accounts in <30secs
using a local mainnet replica.

### Quick Summary of the Ledger Replay process after this Change

1. Find latest applicable account storage (as recent as possible
considering the latest ledger
slot that was saved to disk)
2. Hydrate the bank with all accounts from the storage and return the
slot at which full ledger
replay should continue
3. Replay the ledger considering only blockhashes from the
`full-replay-slot - max-age`
- this is needed since transactions may refer to one of them as the
`most_recent_blockhash`
4. Replay the remaining slots including transactions, i.e. from the
`full-replay-slot` until no
more blocks are found in the ledger
5. Iterate through all accounts in the bank and do the following
depending on the type of
account:

Accounts delegated to us are fetched to update the delegation metadata,
however since our
validator holds the most recent state we don't re-clone them into the
bank.

Non-executable accounts delegated to another validator or not delegated
at all are fully
cloned, including the metadata and are updated in the bank.

Executable program accounts are cloned and updated in the bank the same
way programs are
usually cloned. Since at this point cloning programs already includes
fetching + loading its executable
data accounts we do not explicitly clone those, but ignore them.

## Details

- storage loader improved to work for all files
- also supports max slot in order to load the most recent account state
that is still before
the last ledger entry that was saved to disk
- updated the storage implementation to be able to disable file deletion
after we load it
- added a test specifically for account hydration
- fixed the feature setup to make the tokio-console useable, this was
used to triage issues

### Tooling

#### ledger-stats

The following commands were added/updated:

##### accounts/account

Improved to print more info and added the ability to get logs for a
particular account.

##### log

Added the ability to print logs for a particular account.

##### blockhash

Added to print blockhashses and slots (at this point the latest).

#### genx

This is a tool to be used to generate configs/scripts, etc.
At this point it is used to generate a test-validator script that makes
pre-load accounts from
chain as follows:

1. Find all accounts stored in the ledger
2. Add them to the script as `--maybe-clone` to have them loaded into
the validator on startup
if they exist on chain

This script was used to run a local mainnet replica with all accounts we
need to clone into our
validator pre-loaded. This way developing/triaging the clone hydration
was much faster.

### Unrelated

- fixed the clock value logs even on `trace` to log only at a particular
frequency.
- account dumper now also uses `SanitizedTransactions` and the main
entry path to execute
transactions which means it part-takes in the transaction lock mechanism
to avoid race
conditions
- no longer cloning IDL accounts if they hold 0 lamports

### Remaining

- the account cloning concurrency is hardcoded for now and should be
made configurable in a
follow-up PR [tracked
here](https://github.com/magicblock-labs/magicblock-validator/issues/275)